Fix text output of q-ranker (was labeled as percolator. Also fixed 64-bit compiler warnings in percolator and q-ramker. Code review by Barbara

+++ src/c/q-ranker/DataSet.cpp
@@ -672,7 +672,7 @@ void DataSet::readSQT(const string fname, IntraSetRelation * intraRel,const stri
   string seq;
   
   fileId = fname;
-  unsigned int spos = fileId.rfind('/');
+  size_t spos = fileId.rfind('/');
   if (spos!=string::npos)
     fileId.erase(0,spos+1);
   spos = fileId.find('.');
